app/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-datastore==2.20.1

